[ADD] validacao para caso o usuario queira acessar o arquivo de removerPublicacao pela url

+++ JSP/avaliacao_jsp/src/main/webapp/removerPublicacao.jsp
@@ -11,25 +11,36 @@
 <body>
 	<%
 	
-		int codigo = Integer.parseInt(request.getParameter("codigo"));
-		
-		Conexao c = new Conexao();
-		
-		String deleteCometario = "DELETE FROM comentario WHERE codigo_publicacao = ?";
-		
-		PreparedStatement pstmt = c.efetuarConexao().prepareStatement(deleteCometario);
-		pstmt.setInt(1, codigo);
-		
-		pstmt.execute();
-		
-		String deletePublicacao = "DELETE FROM publicacao WHERE codigo = ?";
-		
-		pstmt = c.efetuarConexao().prepareStatement(deletePublicacao);
-		pstmt.setInt(1, codigo);
-		
-		pstmt.execute();
-		
-		response.sendRedirect("index.jsp");
+		String email=(String)session.getAttribute("email");
+	
+		if (email != null) {
+			try {
+				int codigo = Integer.parseInt(request.getParameter("codigo"));
+				
+				Conexao c = new Conexao();
+				
+				String deleteCometario = "DELETE FROM comentario WHERE codigo_publicacao = ?";
+				
+				PreparedStatement pstmt = c.efetuarConexao().prepareStatement(deleteCometario);
+				pstmt.setInt(1, codigo);
+				
+				pstmt.execute();
+				
+				String deletePublicacao = "DELETE FROM publicacao WHERE codigo = ?";
+				
+				pstmt = c.efetuarConexao().prepareStatement(deletePublicacao);
+				pstmt.setInt(1, codigo);
+				
+				pstmt.execute();
+				
+				response.sendRedirect("index.jsp");
+			} catch(Exception e) {
+				response.sendRedirect("index.jsp");
+			}
+		} else {
+			response.sendRedirect("index.jsp");
+		}
+
 		
 	%>
 </body>
